Fix for RMAN-22201
Refreshing PxrOSL nodes was failing

* double check the length of color parameters. The length for the default_value for our color sockets is 4, so make sure to
increase/decrease the length when necessary
* cannot deepcopy Blender's bpy_prop_array, so manually create a
list and copy the values.
* also bump the addon version and add a changelog for 26.1.

+++ rman_bl_nodes/rman_bl_nodes_sockets.py
@@ -436,7 +436,7 @@ class RendermanNodeSocketFloat(NodeSocketFloat, RendermanSocket):
     renderman_type: StringProperty(default='float')
     is_array: BoolProperty(default=False)
     array_elem: IntProperty(default=-1)
-    default_value: IntProperty(update=update_func)
+    default_value: FloatProperty(update=update_func)
         
     @classmethod
     def draw_color_simple(cls):
